+++ autode/wrappers/keywords.py
@@ -1,5 +1,6 @@
 from typing import Union
 from copy import deepcopy
+from abc import ABC, abstractmethod
 from autode.log import logger
 
 
@@ -328,7 +329,11 @@ class SinglePointKeywords(Keywords):
         return self._string(prefix='SPKeywords')
 
 
-class Keyword:
+class Keyword(ABC):
+
+    @abstractmethod
+    def __repr__(self):
+        """Representation of this keyword"""
 
     def __eq__(self, other):
         return str(self) == str(other)
@@ -402,13 +407,22 @@ class Keyword:
 class BasisSet(Keyword):
     """Basis set for a QM method"""
 
+    def __repr__(self):
+        return f'BasisSet({self.name})'
+
 
 class DispersionCorrection(Keyword):
     """Functional for a DFT method"""
 
+    def __repr__(self):
+        return f'DispersionCorrection({self.name})'
+
 
 class Functional(Keyword):
     """Functional for a DFT method"""
+
+    def __repr__(self):
+        return f'Functional({self.name})'
 
 
 class ImplicitSolventType(Keyword):
@@ -418,17 +432,29 @@ class ImplicitSolventType(Keyword):
         cpcm = ImplicitSolventType(name='cpcm', doi='10.the_doi')
     """
 
+    def __repr__(self):
+        return f'ImplicitSolventType({self.name})'
+
 
 class RI(Keyword):
     """Resolution of identity approximation"""
+
+    def __repr__(self):
+        return f'ResolutionOfIdentity({self.name})'
 
 
 class WFMethod(Keyword):
     """Keyword for a wavefunction method e.g. HF or CCSD(T)"""
 
+    def __repr__(self):
+        return f'WaveFunctionMethod({self.name})'
+
 
 class ECP(Keyword):
     """Effective core potential"""
+
+    def __repr__(self):
+        return f'EffectiveCorePotential({self.name})'
 
     def __eq__(self, other):
         """Equality of ECPs"""
@@ -456,6 +482,9 @@ class ECP(Keyword):
 
 class MaxOptCycles(Keyword):
     """Maximum number of optimisation cycles"""
+
+    def __repr__(self):
+        return f'MaxOptCycles(N = {self.name})'
 
     def __int__(self):
         return int(self.name)