+++ doc/conf.py
@@ -16,7 +16,6 @@ version = autode.__version__
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = ['sphinx.ext.autodoc',
-              'sphinx_rtd_theme',
               'sphinx.ext.napoleon']
 
 # Add any paths that contain templates here, relative to this directory.