+++ doc/changelog.rst
@@ -12,6 +12,11 @@ Usability improvements/Changes
 - Adds `typing <https://docs.python.org/3/library/typing.html>`_ to user-facing functions
 - Updates documentation for readability
 
+Functionality improvements
+**************************
+
+- Changes :code:`Keyword` to an abstract base class
+
 
 1.1.0
 --------